Slim v3: PHP 8.1 - 8.5 support

+++ tests/RouterTest.php
@@ -8,13 +8,13 @@
 namespace Slim\Tests;
 
 use InvalidArgumentException;
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 use ReflectionClass;
 use RuntimeException;
 use Slim\Http\Uri;
 use Slim\Router;
 
-class RouterTest extends PHPUnit_Framework_TestCase
+class RouterTest extends TestCase
 {
     /**
      * @var Router
@@ -26,14 +26,14 @@ class RouterTest extends PHPUnit_Framework_TestCase
      */
     protected $cacheFile;
 
-    public function setUp()
+    public function setUp(): void
     {
         $this->router = new Router;
     }
 
-    public function tearDown()
+    public function tearDown(): void
     {
-        if (file_exists($this->cacheFile)) {
+        if ($this->cacheFile && file_exists($this->cacheFile)) {
             unlink($this->cacheFile);
         }
     }
@@ -48,7 +48,7 @@ class RouterTest extends PHPUnit_Framework_TestCase
         $route = $this->router->map($methods, $pattern, $callable);
 
         $this->assertInstanceOf('\Slim\Interfaces\RouteInterface', $route);
-        $this->assertAttributeContains($route, 'routes', $this->router);
+        $this->assertContains($route, $this->router->getRoutes());
     }
 
     public function testMapPrependsGroupPattern()
@@ -64,15 +64,13 @@ class RouterTest extends PHPUnit_Framework_TestCase
         $route = $this->router->map($methods, $pattern, $callable);
         $this->router->popGroup();
 
-        $this->assertAttributeEquals('/prefix/hello/{first}/{last}', 'pattern', $route);
+        $this->assertSame('/prefix/hello/{first}/{last}', $route->getPattern());
     }
 
-    /**
-     * @expectedException InvalidArgumentException
-     * @expectedExceptionMessage Route pattern must be a string
-     */
     public function testMapWithInvalidPatternType()
     {
+        $this->expectException(\InvalidArgumentException::class);
+        $this->expectExceptionMessage('Route pattern must be a string');
         $methods = ['GET'];
         $pattern = ['foo'];
         $callable = function ($request, $response, $args) {
@@ -196,11 +194,9 @@ class RouterTest extends PHPUnit_Framework_TestCase
         );
     }
 
-    /**
-     * @expectedException InvalidArgumentException
-     */
     public function testPathForWithMissingSegmentData()
     {
+        $this->expectException(\InvalidArgumentException::class);
         $methods = ['GET'];
         $pattern = '/hello/{first}/{last}';
         $callable = function ($request, $response, $args) {
@@ -212,11 +208,9 @@ class RouterTest extends PHPUnit_Framework_TestCase
         $this->router->pathFor('foo', ['last' => 'lockhart']);
     }
 
-    /**
-     * @expectedException RuntimeException
-     */
     public function testPathForRouteNotExists()
     {
+        $this->expectException(\RuntimeException::class);
         $methods = ['GET'];
         $pattern = '/hello/{first}/{last}';
         $callable = function ($request, $response, $args) {
@@ -228,11 +222,9 @@ class RouterTest extends PHPUnit_Framework_TestCase
         $this->router->pathFor('bar', ['first' => 'josh', 'last' => 'lockhart']);
     }
 
-    /**
-     * @expectedException InvalidArgumentException
-     */
     public function testSettingInvalidBasePath()
     {
+        $this->expectException(\InvalidArgumentException::class);
         $this->router->setBasePath(['invalid']);
     }
 
@@ -240,7 +232,6 @@ class RouterTest extends PHPUnit_Framework_TestCase
     {
         $class = new ReflectionClass($this->router);
         $method = $class->getMethod('createDispatcher');
-        $method->setAccessible(true);
         $this->assertInstanceOf('\FastRoute\Dispatcher', $method->invoke($this->router));
     }
 
@@ -252,15 +243,12 @@ class RouterTest extends PHPUnit_Framework_TestCase
         }));
         $class = new ReflectionClass($this->router);
         $prop = $class->getProperty('dispatcher');
-        $prop->setAccessible(true);
         $this->assertInstanceOf('\FastRoute\Dispatcher', $prop->getValue($this->router));
     }
 
-    /**
-     * @expectedException RuntimeException
-     */
     public function testRemoveRoute()
     {
+        $this->expectException(\RuntimeException::class);
         $methods = ['GET'];
         $callable = function ($request, $response, $args) {
             echo sprintf('Hello ignore me');
@@ -312,11 +300,9 @@ class RouterTest extends PHPUnit_Framework_TestCase
         $this->router->getNamedRoute($routeToRemove->getName());
     }
 
-    /**
-     * @expectedException RuntimeException
-     */
     public function testRouteRemovalNotExists()
     {
+        $this->expectException(\RuntimeException::class);
         $this->router->setBasePath('/base/path');
         $this->router->removeNamedRoute('non-existing-route-name');
     }
@@ -347,17 +333,14 @@ class RouterTest extends PHPUnit_Framework_TestCase
 
         $class = new ReflectionClass($this->router);
         $property = $class->getProperty('cacheFile');
-        $property->setAccessible(true);
 
         $this->assertFalse($property->getValue($this->router));
     }
 
     public function testSettingInvalidCacheFileValue()
     {
-        $this->setExpectedException(
-            '\InvalidArgumentException',
-            'Router cache file must be a string'
-        );
+        $this->expectException(\InvalidArgumentException::class);
+        $this->expectExceptionMessage('Router cache file must be a string');
         $this->router->setCacheFile(['invalid']);
     }
 
@@ -366,10 +349,8 @@ class RouterTest extends PHPUnit_Framework_TestCase
         $this->cacheFile = __DIR__ . '/non-readable.cache';
         file_put_contents($this->cacheFile, '<?php return []; ?>');
 
-        $this->setExpectedException(
-            '\RuntimeException',
-            sprintf('Router cache file `%s` is not readable', $this->cacheFile)
-        );
+        $this->expectException(\RuntimeException::class);
+        $this->expectExceptionMessage(sprintf('Router cache file `%s` is not readable', $this->cacheFile));
 
         $this->router->setCacheFile($this->cacheFile);
     }
@@ -378,10 +359,8 @@ class RouterTest extends PHPUnit_Framework_TestCase
     {
         $cacheFile = __DIR__ . '/non-writable-directory/router.cache';
 
-        $this->setExpectedException(
-            '\RuntimeException',
-            sprintf('Router cache file directory `%s` is not writable', dirname($cacheFile))
-        );
+        $this->expectException(\RuntimeException::class);
+        $this->expectExceptionMessage(sprintf('Router cache file directory `%s` is not writable', dirname($cacheFile)));
 
         $this->router->setCacheFile($cacheFile);
     }
@@ -399,7 +378,6 @@ class RouterTest extends PHPUnit_Framework_TestCase
         $this->router->setCacheFile($cacheFile);
         $class = new ReflectionClass($this->router);
         $method = $class->getMethod('createDispatcher');
-        $method->setAccessible(true);
 
         $dispatcher = $method->invoke($this->router);
         $this->assertInstanceOf('\FastRoute\Dispatcher', $dispatcher);
@@ -412,7 +390,6 @@ class RouterTest extends PHPUnit_Framework_TestCase
 
         $class = new ReflectionClass($router2);
         $method = $class->getMethod('createDispatcher');
-        $method->setAccessible(true);
 
         $dispatcher2 = $method->invoke($this->router);
         $result = $dispatcher2->dispatch('GET', '/hello/josh/lockhart');
@@ -429,18 +406,15 @@ class RouterTest extends PHPUnit_Framework_TestCase
     {
         $class = new ReflectionClass($this->router);
         $method = $class->getMethod('createDispatcher');
-        $method->setAccessible(true);
 
         $dispatcher = $method->invoke($this->router);
         $dispatcher2 = $method->invoke($this->router);
         $this->assertSame($dispatcher2, $dispatcher);
     }
 
-    /**
-     * @expectedException RuntimeException
-     */
     public function testLookupRouteThrowsExceptionIfRouteNotFound()
     {
+        $this->expectException(\RuntimeException::class);
         $this->router->lookupRoute("thisIsMissing");
     }
